+++ NEWS.md
@@ -1,3 +1,7 @@
+# *arg* (development version)
+
+* Fixed bugs with `arg_or()`, `arg_and()`, `when_supplied()`, and `when_not_null()`.
+
 # *arg* 0.1.0
 
 * Initial CRAN submission.